Number of Children
Class metric : Number of Children 완료

+++ src/main/java/net/sourceforge/pmd/lang/java/rule/stanly/metrics/ClassMetric.java
@@ -10,7 +10,7 @@ public class ClassMetric {
 	private int EfferentCoupling;
 	private float WMC;
 	private int DIT;
-	private float NOC;
+	private int NOC;
 	private float CBO;
 	private float RFC;
 	private float LCOM;
@@ -72,10 +72,10 @@ public class ClassMetric {
 	public void setDIT(int dIT) {
 		DIT = dIT;
 	}
-	public float getNOC() {
+	public int getNOC() {
 		return NOC;
 	}
-	public void setNOC(float nOC) {
+	public void setNOC(int nOC) {
 		NOC = nOC;
 	}
 	public float getCBO() {
